Move query from user util to favor util
