+++ app/src/main/java/ch/epfl/favo/favor/FavorUtil.java
@@ -158,4 +158,12 @@ public class FavorUtil {
     collection.updateDocument(favor.getId(), updates);
     favor.setPictureUrl(url);
   }
+
+  public Query getAllUserFavors(String userId) {
+    return collection
+            .getReference()
+            .orderBy("title", Query.Direction.ASCENDING)
+            .orderBy("postedTime", Query.Direction.DESCENDING)
+            .whereArrayContains("userIds", userId);
+  }
 }
