+++ app/src/main/java/ch/epfl/favo/view/tabs/favorList/FavorPage.java
@@ -31,7 +31,7 @@ import com.google.firebase.firestore.Query;
 import ch.epfl.favo.MainActivity;
 import ch.epfl.favo.R;
 import ch.epfl.favo.favor.Favor;
-import ch.epfl.favo.user.UserUtil;
+import ch.epfl.favo.favor.FavorUtil;
 import ch.epfl.favo.util.DependencyFactory;
 
 import static ch.epfl.favo.util.CommonTools.hideSoftKeyboard;
@@ -98,7 +98,7 @@ public class FavorPage extends Fragment {
     mSwipeRefreshLayout = rootView.findViewById(R.id.swipe_refresh_layout);
 
     baseQuery =
-        UserUtil.getSingleInstance()
+        FavorUtil.getSingleInstance()
             .getAllUserFavors(DependencyFactory.getCurrentFirebaseUser().getUid());
 
     activeFavorsOptions = createFirestorePagingOptions(baseQuery.whereEqualTo("isArchived", false));
