+++ app/src/test/java/ch/epfl/favo/user/UserUtilTest.java
@@ -153,8 +153,5 @@ public class UserUtilTest {
   public void testUserReference(){
     UserUtil.getSingleInstance().getCurrentUserReference("randomId");
   }
-  @Test
-  public void testActiveFavorQuery(){
-    UserUtil.getSingleInstance().getAllUserFavors("randomId");
-  }
+
 }
